Tetros class moved to Game.py
